+++ pyproject.toml
@@ -84,6 +84,9 @@ auth-tool = "asreview.webapp.entry_points.auth_tool:AuthTool"
 simulate = "asreview.simulation.cli:_cli_simulate"
 algorithms = "asreview.models.cli:cli_algorithms"
 
+[project.entry-points."asreview.entry_points_internal"]
+task-manager = "asreview.webapp.entry_points.task_manager:main"
+
 [project.entry-points."asreview.readers"]
 ".csv" = "asreview.data:CSVReader"
 ".tab" = "asreview.data:CSVReader"
